refactor: move ltc decoder to module
